fix label tests #2

+++ src/main/java/hexlet/code/service/LabelService.java
@@ -3,14 +3,16 @@ package hexlet.code.service;
 import hexlet.code.dto.LabelCreateDTO;
 import hexlet.code.dto.LabelDTO;
 import hexlet.code.dto.LabelUpdateDTO;
+import hexlet.code.exception.ConflictException;
 import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.mapper.LabelMapper;
 import hexlet.code.repository.LabelRepository;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 @Service
 @Slf4j
@@ -87,6 +89,15 @@ public class LabelService {
                                         id
                                 )
                         ));
+
+        if (!label.getTasks().isEmpty()) {
+            throw new ConflictException(
+                    String.format(
+                            "Label with id %d is used by tasks",
+                            id
+                    )
+            );
+        }
 
         labelRepository.delete(label);
     }
